requirements.txt
asgiref==3.2.7
astroid==2.4.1
Django==3.0.6
django-tinymce==3.0.2
djangorestframework==3.11.0
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
